The excision: retire family rt and its hand-written shape code (#196)
Per the owner ruling on #196. The generated subject via the data-driven
bench (#191/#197) is the estate's only benchmark; the hand-written
runtime-usage product retires with the nine repos' in-repo benches.

Family bits shared files with rt in go, rust and cs — those files are
split, not deleted (bits.go, bits.rs, BitsBench.cs). The Bench corpus,
its goldens and the §1.5 oracle gate stay: they serve the gen family.
BENCH-STANDARD §1.8's bench_wstring row is marked SUSPENDED — schema
defers wstring, so no generated codec can carry it and the family that
would have is gone.

The bench/LOCK reproduction control was run: C reproduces within 2.6%,
C++ bench_packet read and bench_bits read shift -11% and -9.6% and the
shifts repeat against a stable window. Byte-identical library and
generated code, so this is harness code-layout sensitivity — a
measurement-continuity break needing a #194 sub-era boundary, not a
performance change. Detail in the PR body.
