+++ bench/c/bench_main.c
@@ -39,7 +39,7 @@
 #include "TypesWire.h"
 #include "WireWire.h"
 #include "RealWorldWire.h"      /* generated/bench/c — the §1.7 realistic snapshot (real_packet) */
-#include "BenchWire.h"          /* generated/bench/c — the Bench corpus GENERATED (the gen twins of the rt rows, issue #177) */
+#include "BenchWire.h"          /* generated/bench/c — the Bench corpus GENERATED (issue #177) */
 
 static volatile uint64_t g_sink = 0;    /* defeats dead code elimination of computed values */
 
@@ -109,7 +109,7 @@ static const char * g_variant_dir = "bench/corpus/variants";
 #ifndef BENCH_OPT
 #define BENCH_OPT "O3"
 #endif
-/* family is per ROW now (gen | rt | bits — §5.1); linkage/checks/opt/inline
+/* family is per ROW now (gen | bits — §5.1); linkage/checks/opt/inline
    stay per-runner constants */
 #define CsvSuffix "hdr,removed," BENCH_OPT ",unknown"
 #define MaxCsvRows 64
@@ -630,13 +630,11 @@ static void vary_real_packet( RealPacket * m, uint64_t rng )
 
 /* ------------------------------------------------------------------------------------------
    family gen over the Bench corpus (issue #177): the four Bench.schema shapes
-   measured through the GENERATED code (generated/bench/c/BenchWire.h) — the
-   gen twins of the rt rows below, which serialize the same shapes BY HAND
-   against the runtime API. Same golden files, same pinned values, same LCG
-   field mappings (vary_rt_* on the generated structs), same bench_message.inc
-   discipline as every gen row above; the family column carries the subject,
-   and relative.go refuses gen-vs-rt ratios. Generated best case per the
-   profiling doctrine (#170): the plain optimized release build, no PGO.
+   measured through the GENERATED code (generated/bench/c/BenchWire.h) — same
+   golden files, same pinned values, same LCG field mappings, same
+   bench_message.inc discipline as every gen row above. Generated best case
+   per the profiling doctrine (#170): the plain optimized release build, no
+   PGO.
    ------------------------------------------------------------------------------------------ */
 
 static void vary_gen_packet( BenchPacket * p, uint64_t rng )
@@ -843,649 +841,7 @@ static TestData pin_testdata( void )
 }
 
 
-/* ------------------------------------------------------------------------------------------
-   family rt (BENCH-STANDARD.md §1.3, §1.5): the serialize.c runtime API
-   called BY HAND — the four Bench.schema shapes as hand-written packets, the
-   way serialize.c/bench.c writes them. The §1.5 oracle gate byte-compares the
-   hand-written wire against the goldens the GENERATED code pinned
-   (testdata/wire/bench_*.bin) and round-trips before any number. Internal
-   linkage per §3.1 (static). Per §3.2 every benched op has EXACTLY two call
-   sites: the untimed once-helper and its timed loop.
-   ------------------------------------------------------------------------------------------ */
-
 #define BENCH_NOINLINE __attribute__(( noinline ))
-
-typedef struct RtBenchPacket
-{
-    serialize_int32_t a, b, c;
-    serialize_uint32_t bits7, bits13, bits23;
-    int flag;
-    float x, y, z;
-    serialize_uint64_t big;
-    serialize_uint8_t blob[17];
-} RtBenchPacket;
-
-static int rt_write_bench_packet( serialize_write_stream_t * stream, const RtBenchPacket * p )
-{
-    if ( !serialize_write_int( stream, p->a, -100, +100 ) ) { return 0; }
-    if ( !serialize_write_int( stream, p->b, 0, 65535 ) ) { return 0; }
-    if ( !serialize_write_int( stream, p->c, -1000000, +1000000 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, p->bits7, 7 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, p->bits13, 13 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, p->bits23, 23 ) ) { return 0; }
-    if ( !serialize_write_bool( stream, p->flag ) ) { return 0; }
-    if ( !serialize_write_float( stream, p->x ) ) { return 0; }
-    if ( !serialize_write_float( stream, p->y ) ) { return 0; }
-    if ( !serialize_write_float( stream, p->z ) ) { return 0; }
-    if ( !serialize_write_uint64( stream, p->big ) ) { return 0; }
-    if ( !serialize_write_bytes( stream, p->blob, (int) sizeof( p->blob ) ) ) { return 0; }  /* aligns internally */
-    return 1;
-}
-
-static int rt_read_bench_packet( serialize_read_stream_t * stream, RtBenchPacket * p )
-{
-    if ( !serialize_read_int( stream, &p->a, -100, +100 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &p->b, 0, 65535 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &p->c, -1000000, +1000000 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &p->bits7, 7 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &p->bits13, 13 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &p->bits23, 23 ) ) { return 0; }
-    if ( !serialize_read_bool( stream, &p->flag ) ) { return 0; }
-    if ( !serialize_read_float( stream, &p->x ) ) { return 0; }
-    if ( !serialize_read_float( stream, &p->y ) ) { return 0; }
-    if ( !serialize_read_float( stream, &p->z ) ) { return 0; }
-    if ( !serialize_read_uint64( stream, &p->big ) ) { return 0; }
-    if ( !serialize_read_bytes( stream, p->blob, (int) sizeof( p->blob ) ) ) { return 0; }
-    return 1;
-}
-
-typedef struct RtBenchInts
-{
-    serialize_int32_t f0, f1, f2, f3, f4, f5, f6, f7, f8, f9;
-} RtBenchInts;
-
-static int rt_write_bench_ints( serialize_write_stream_t * stream, const RtBenchInts * f )
-{
-    if ( !serialize_write_int( stream, f->f0, -100, +100 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f1, 0, 65535 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f2, -1000000, +1000000 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f3, 0, 3 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f4, -15, +15 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f5, 0, 1000 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f6, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f7, 0, 255 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f8, -600000, +600000 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->f9, 0, 100 ) ) { return 0; }
-    return 1;
-}
-
-static int rt_read_bench_ints( serialize_read_stream_t * stream, RtBenchInts * f )
-{
-    if ( !serialize_read_int( stream, &f->f0, -100, +100 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f1, 0, 65535 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f2, -1000000, +1000000 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f3, 0, 3 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f4, -15, +15 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f5, 0, 1000 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f6, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f7, 0, 255 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f8, -600000, +600000 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->f9, 0, 100 ) ) { return 0; }
-    return 1;
-}
-
-typedef struct RtBenchBits
-{
-    serialize_uint32_t b7, b13, b23, b3, b32, b11, b19;
-    serialize_uint64_t b48;
-} RtBenchBits;
-
-/* 48 bits ride as lo32 + hi16, low first — the wire order the generated C
-   emits and serialize.c's 64-bit convention */
-static int rt_write_bench_bits( serialize_write_stream_t * stream, const RtBenchBits * f )
-{
-    if ( !serialize_write_bits( stream, f->b7, 7 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b13, 13 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b23, 23 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b3, 3 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b32, 32 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b11, 11 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->b19, 19 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->b48 & 0xFFFFFFFFu ), 32 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->b48 >> 32 ), 16 ) ) { return 0; }
-    return 1;
-}
-
-static int rt_read_bench_bits( serialize_read_stream_t * stream, RtBenchBits * f )
-{
-    serialize_uint32_t lo, hi;
-    if ( !serialize_read_bits( stream, &f->b7, 7 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b13, 13 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b23, 23 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b3, 3 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b32, 32 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b11, 11 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->b19, 19 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &lo, 32 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &hi, 16 ) ) { return 0; }
-    f->b48 = (serialize_uint64_t) lo | ( (serialize_uint64_t) hi << 32 );
-    return 1;
-}
-
-/* BenchMixed by hand (issue #184): every serialize_* stream operation the
-   schema language expresses, in the order write_bench_mixed emits them. The
-   §1.5 oracle gate byte-compares this against the generated code's golden. */
-typedef struct RtMixedEntity
-{
-    serialize_uint32_t entity_id;
-    serialize_int32_t pos_x, pos_y, pos_z;
-    serialize_uint32_t yaw, pitch;
-    serialize_int32_t vel_x, vel_y, vel_z;
-    serialize_int32_t health;
-    serialize_int32_t weapon;      /* the enum wire */
-    serialize_uint32_t damage;     /* the flags wire, 8 bits */
-    int moving, firing;
-} RtMixedEntity;
-
-typedef struct RtMixedStat
-{
-    serialize_uint32_t stat_id;
-    serialize_int32_t delta;
-} RtMixedStat;
-
-typedef struct RtMixedHitEvent
-{
-    serialize_uint32_t target_id;
-    serialize_int32_t damage, hit_kind;
-    int crit;
-} RtMixedHitEvent;
-
-typedef struct RtMixedChatEvent
-{
-    serialize_int32_t channel;
-    serialize_uint32_t speaker;
-} RtMixedChatEvent;
-
-typedef struct RtMixedPickupEvent
-{
-    serialize_uint32_t item_id;
-    serialize_int32_t amount;
-} RtMixedPickupEvent;
-
-typedef struct RtBenchMixed
-{
-    serialize_uint32_t magic;
-    serialize_uint32_t sequence;
-    serialize_int32_t ack_sequence;
-    serialize_uint32_t ack_bits;
-    serialize_uint64_t session_id;
-    serialize_uint32_t client_id;
-    serialize_uint64_t nonce;
-    serialize_int64_t world_time;
-    serialize_uint64_t frame_tick;
-    serialize_int32_t server_time;          /* raw Q24.8 */
-    serialize_int32_t entities_count;
-    RtMixedEntity entities[8];
-    serialize_int32_t stats_count;
-    RtMixedStat stats[80];
-    serialize_int32_t event_type;           /* the union tag: 0 = None */
-    RtMixedHitEvent hit;
-    RtMixedChatEvent chat;
-    RtMixedPickupEvent pickup;
-    serialize_uint8_t loadout[4];
-    serialize_int32_t player_name_length;
-    char player_name[16];
-    serialize_int32_t payload_length;
-    serialize_uint8_t payload[16];
-    float aim_x, aim_y, aim_z;
-    float recoil;
-    double drift;
-    serialize_uint128_t wide_key;
-    serialize_int128_t flux;
-    serialize_uint16_t ping;                /* raw UQ8.8 */
-    serialize_uint32_t reserved_bits;
-    serialize_uint32_t crc_hint;
-    int has_extra;
-    serialize_int32_t extra, idle_ticks;
-} RtBenchMixed;
-
-/* the +/-2^100 band flux rides in, composed the way serialize.c spells 128-bit
-   constants (no C literal exists) */
-#define RT_FLUX_MIN serialize_int128_make( 18446744004990074880ULL, 0ULL )
-#define RT_FLUX_MAX serialize_int128_make( 68719476736ULL, 0ULL )
-
-static int rt_write_mixed_entity( serialize_write_stream_t * stream, const RtMixedEntity * e )
-{
-    if ( !serialize_write_bits( stream, e->entity_id, 12 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->pos_x, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->pos_y, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->pos_z, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, e->yaw, 9 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, e->pitch, 9 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->vel_x, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->vel_y, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->vel_z, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->health, 0, 1000 ) ) { return 0; }
-    if ( !serialize_write_int( stream, e->weapon, 0, 15 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, e->damage, 8 ) ) { return 0; }
-    if ( !serialize_write_bool( stream, e->moving ) ) { return 0; }
-    if ( !serialize_write_bool( stream, e->firing ) ) { return 0; }
-    return 1;
-}
-
-static int rt_read_mixed_entity( serialize_read_stream_t * stream, RtMixedEntity * e )
-{
-    if ( !serialize_read_bits( stream, &e->entity_id, 12 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->pos_x, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->pos_y, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->pos_z, -16383, +16383 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &e->yaw, 9 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &e->pitch, 9 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->vel_x, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->vel_y, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->vel_z, -2048, +2047 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->health, 0, 1000 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &e->weapon, 0, 15 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &e->damage, 8 ) ) { return 0; }
-    if ( !serialize_read_bool( stream, &e->moving ) ) { return 0; }
-    if ( !serialize_read_bool( stream, &e->firing ) ) { return 0; }
-    return 1;
-}
-
-static int rt_write_bench_mixed( serialize_write_stream_t * stream, const RtBenchMixed * f )
-{
-    int i;
-    serialize_int64_t ping_fixed;
-    if ( !serialize_write_bits( stream, f->magic, 16 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->sequence, 16 ) ) { return 0; }
-    if ( !serialize_write_int( stream, f->ack_sequence, 0, 65535 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->ack_bits, 32 ) ) { return 0; }
-    if ( !serialize_write_uint64( stream, f->session_id ) ) { return 0; }
-    if ( !serialize_write_uint32( stream, f->client_id ) ) { return 0; }
-    /* the full-unsigned ranged path is width-computed raw bits, low 32 first */
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->nonce & 0xFFFFFFFFu ), 32 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->nonce >> 32 ), 32 ) ) { return 0; }
-    if ( !serialize_write_int64( stream, f->world_time, -1000000000000LL, 1000000000000LL ) ) { return 0; }
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->frame_tick & 0xFFFFFFFFu ), 32 ) ) { return 0; }
-    if ( !serialize_write_bits( stream, (serialize_uint32_t) ( f->frame_tick >> 32 ), 16 ) ) { return 0; }
-    if ( !serialize_write_fixed32( stream, f->server_time, 24, 8, 0, 65535 ) ) { return 0; }
-
-    if ( !serialize_write_int( stream, f->entities_count, 1, 8 ) ) { return 0; }
-    for ( i = 0; i < f->entities_count; i++ )
-        if ( !rt_write_mixed_entity( stream, &f->entities[i] ) ) { return 0; }
-
-    if ( !serialize_write_int( stream, f->stats_count, 0, 80 ) ) { return 0; }
-    for ( i = 0; i < f->stats_count; i++ )
-    {
-        if ( !serialize_write_bits( stream, f->stats[i].stat_id, 8 ) ) { return 0; }
-        if ( !serialize_write_int( stream, f->stats[i].delta, -512, +511 ) ) { return 0; }
-    }
-
-    if ( !serialize_write_int( stream, f->event_type, 0, 3 ) ) { return 0; }
-    if ( f->event_type == 1 )
-    {
-        if ( !serialize_write_bits( stream, f->hit.target_id, 12 ) ) { return 0; }
-        if ( !serialize_write_int( stream, f->hit.damage, 0, 4095 ) ) { return 0; }
-        if ( !serialize_write_int( stream, f->hit.hit_kind, 0, 7 ) ) { return 0; }
-        if ( !serialize_write_bool( stream, f->hit.crit ) ) { return 0; }
-    }
-    else if ( f->event_type == 2 )
-    {
-        if ( !serialize_write_int( stream, f->chat.channel, 0, 3 ) ) { return 0; }
-        if ( !serialize_write_bits( stream, f->chat.speaker, 12 ) ) { return 0; }
-    }
-    else if ( f->event_type == 3 )
-    {
-        if ( !serialize_write_bits( stream, f->pickup.item_id, 10 ) ) { return 0; }
-        if ( !serialize_write_int( stream, f->pickup.amount, 0, 255 ) ) { return 0; }
-    }
-
-    for ( i = 0; i < 4; i++ )
-        if ( !serialize_write_uint8( stream, f->loadout[i] ) ) { return 0; }
-
-    /* string(15) and bytes(16) ride as their §4.3 decomposition in every rt
-       leg — see bench/cpp/bench_main.cpp for the reasoning */
-    if ( !serialize_write_int( stream, f->player_name_length, 0, 15 ) ) { return 0; }
-    if ( !serialize_write_bytes( stream, (const serialize_uint8_t *) f->player_name, (int) f->player_name_length ) ) { return 0; }
-
-    if ( !serialize_write_int( stream, f->payload_length, 0, 16 ) ) { return 0; }
-    if ( !serialize_write_bytes( stream, f->payload, (int) f->payload_length ) ) { return 0; }
-
-    if ( !serialize_write_compressed_float( stream, f->aim_x, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_write_compressed_float( stream, f->aim_y, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_write_compressed_float( stream, f->aim_z, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_write_float( stream, f->recoil ) ) { return 0; }
-    if ( !serialize_write_double( stream, f->drift ) ) { return 0; }
-    if ( !serialize_write_uint128( stream, f->wide_key ) ) { return 0; }
-    if ( !serialize_write_int128( stream, f->flux, RT_FLUX_MIN, RT_FLUX_MAX ) ) { return 0; }
-    ping_fixed = (serialize_int64_t) f->ping;
-    if ( !serialize_write_fixed64( stream, ping_fixed, 8, 8, 0, 250 ) ) { return 0; }
-
-    if ( !serialize_write_bits( stream, f->reserved_bits, 4 ) ) { return 0; }
-    if ( !serialize_write_align( stream ) ) { return 0; }
-    if ( !serialize_write_bits( stream, f->crc_hint, 24 ) ) { return 0; }
-    if ( !serialize_write_bool( stream, f->has_extra ) ) { return 0; }
-    if ( f->has_extra )
-    {
-        if ( !serialize_write_int( stream, f->extra, 0, 255 ) ) { return 0; }
-    }
-    else
-    {
-        if ( !serialize_write_int( stream, f->idle_ticks, 0, 15 ) ) { return 0; }
-    }
-    return 1;
-}
-
-static int rt_read_bench_mixed( serialize_read_stream_t * stream, RtBenchMixed * f )
-{
-    int i;
-    serialize_uint32_t lo, hi;
-    serialize_int64_t ping_fixed;
-    if ( !serialize_read_bits( stream, &f->magic, 16 ) ) { return 0; }
-    if ( f->magic != 0xC0DE ) { return 0; }   /* const(0xC0DE, 16): a read REJECTS any other value */
-    if ( !serialize_read_bits( stream, &f->sequence, 16 ) ) { return 0; }
-    if ( !serialize_read_int( stream, &f->ack_sequence, 0, 65535 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->ack_bits, 32 ) ) { return 0; }
-    if ( !serialize_read_uint64( stream, &f->session_id ) ) { return 0; }
-    if ( !serialize_read_uint32( stream, &f->client_id ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &lo, 32 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &hi, 32 ) ) { return 0; }
-    f->nonce = (serialize_uint64_t) lo | ( (serialize_uint64_t) hi << 32 );
-    if ( !serialize_read_int64( stream, &f->world_time, -1000000000000LL, 1000000000000LL ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &lo, 32 ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &hi, 16 ) ) { return 0; }
-    f->frame_tick = (serialize_uint64_t) lo | ( (serialize_uint64_t) hi << 32 );
-    if ( !serialize_read_fixed32( stream, &f->server_time, 24, 8, 0, 65535 ) ) { return 0; }
-
-    if ( !serialize_read_int( stream, &f->entities_count, 1, 8 ) ) { return 0; }
-    for ( i = 0; i < f->entities_count; i++ )
-        if ( !rt_read_mixed_entity( stream, &f->entities[i] ) ) { return 0; }
-
-    if ( !serialize_read_int( stream, &f->stats_count, 0, 80 ) ) { return 0; }
-    for ( i = 0; i < f->stats_count; i++ )
-    {
-        if ( !serialize_read_bits( stream, &f->stats[i].stat_id, 8 ) ) { return 0; }
-        if ( !serialize_read_int( stream, &f->stats[i].delta, -512, +511 ) ) { return 0; }
-    }
-
-    if ( !serialize_read_int( stream, &f->event_type, 0, 3 ) ) { return 0; }
-    if ( f->event_type == 1 )
-    {
-        if ( !serialize_read_bits( stream, &f->hit.target_id, 12 ) ) { return 0; }
-        if ( !serialize_read_int( stream, &f->hit.damage, 0, 4095 ) ) { return 0; }
-        if ( !serialize_read_int( stream, &f->hit.hit_kind, 0, 7 ) ) { return 0; }
-        if ( !serialize_read_bool( stream, &f->hit.crit ) ) { return 0; }
-    }
-    else if ( f->event_type == 2 )
-    {
-        if ( !serialize_read_int( stream, &f->chat.channel, 0, 3 ) ) { return 0; }
-        if ( !serialize_read_bits( stream, &f->chat.speaker, 12 ) ) { return 0; }
-    }
-    else if ( f->event_type == 3 )
-    {
-        if ( !serialize_read_bits( stream, &f->pickup.item_id, 10 ) ) { return 0; }
-        if ( !serialize_read_int( stream, &f->pickup.amount, 0, 255 ) ) { return 0; }
-    }
-
-    for ( i = 0; i < 4; i++ )
-        if ( !serialize_read_uint8( stream, &f->loadout[i] ) ) { return 0; }
-
-    if ( !serialize_read_int( stream, &f->player_name_length, 0, 15 ) ) { return 0; }
-    if ( !serialize_read_bytes( stream, (serialize_uint8_t *) f->player_name, (int) f->player_name_length ) ) { return 0; }
-
-    if ( !serialize_read_int( stream, &f->payload_length, 0, 16 ) ) { return 0; }
-    if ( !serialize_read_bytes( stream, f->payload, (int) f->payload_length ) ) { return 0; }
-
-    if ( !serialize_read_compressed_float( stream, &f->aim_x, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_read_compressed_float( stream, &f->aim_y, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_read_compressed_float( stream, &f->aim_z, -1.0f, 1.0f, 0.01f ) ) { return 0; }
-    if ( !serialize_read_float( stream, &f->recoil ) ) { return 0; }
-    if ( !serialize_read_double( stream, &f->drift ) ) { return 0; }
-    if ( !serialize_read_uint128( stream, &f->wide_key ) ) { return 0; }
-    if ( !serialize_read_int128( stream, &f->flux, RT_FLUX_MIN, RT_FLUX_MAX ) ) { return 0; }
-    if ( !serialize_read_fixed64( stream, &ping_fixed, 8, 8, 0, 250 ) ) { return 0; }
-    f->ping = (serialize_uint16_t) ping_fixed;
-
-    if ( !serialize_read_bits( stream, &f->reserved_bits, 4 ) ) { return 0; }
-    if ( f->reserved_bits != 0 ) { return 0; }   /* reserved(4): a read rejects nonzero */
-    if ( !serialize_read_align( stream ) ) { return 0; }
-    if ( !serialize_read_bits( stream, &f->crc_hint, 24 ) ) { return 0; }
-    if ( !serialize_read_bool( stream, &f->has_extra ) ) { return 0; }
-    if ( f->has_extra )
-    {
-        if ( !serialize_read_int( stream, &f->extra, 0, 255 ) ) { return 0; }
-    }
-    else
-    {
-        if ( !serialize_read_int( stream, &f->idle_ticks, 0, 15 ) ) { return 0; }
-    }
-    return 1;
-}
-
-/* ---- vary functions: bench/cpp/bench_main.cpp's rt mappings exactly ---- */
-
-static void vary_rt_packet( RtBenchPacket * p, uint64_t rng )
-{
-    p->a = (serialize_int32_t) ( ( rng >> 8 ) & 63 ) - 32;
-    p->b = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 16 ) & 65535 );
-    p->c = (serialize_int32_t) ( ( rng >> 24 ) & 0xFFFFF ) - 500000;
-    p->bits7 = (serialize_uint32_t) rng & 127;
-    p->bits13 = (serialize_uint32_t) ( rng >> 3 ) & 8191;
-    p->bits23 = (serialize_uint32_t) ( rng >> 5 ) & 8388607;
-    p->flag = ( rng & 1 ) != 0;
-    p->x = (float) ( (serialize_uint32_t) rng & 0xFFFF );
-    p->big = rng;
-    p->blob[0] = (serialize_uint8_t) ( rng >> 32 );
-}
-
-static void vary_rt_ints( RtBenchInts * f, uint64_t rng )
-{
-    f->f0 = (serialize_int32_t) ( ( rng >> 8 ) & 63 ) - 32;
-    f->f1 = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 16 ) & 65535 );
-    f->f2 = (serialize_int32_t) ( ( rng >> 24 ) & 0xFFFFF ) - 500000;
-    f->f3 = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 2 ) & 3 );
-    f->f4 = (serialize_int32_t) ( ( rng >> 11 ) & 15 ) - 8;
-    f->f5 = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 22 ) & 511 );
-    f->f6 = (serialize_int32_t) ( ( rng >> 33 ) & 2047 ) - 1024;
-    f->f7 = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 40 ) & 255 );
-    f->f8 = (serialize_int32_t) ( ( rng >> 30 ) & 0xFFFFF ) - 500000;
-    f->f9 = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 57 ) & 63 );
-}
-
-static void vary_rt_bits( RtBenchBits * f, uint64_t rng )
-{
-    f->b7 = (serialize_uint32_t) rng & 127;
-    f->b13 = (serialize_uint32_t) ( rng >> 3 ) & 8191;
-    f->b23 = (serialize_uint32_t) ( rng >> 5 ) & 8388607;
-    f->b3 = (serialize_uint32_t) ( rng >> 29 ) & 7;
-    f->b32 = (serialize_uint32_t) ( rng >> 16 );
-    f->b11 = (serialize_uint32_t) ( rng >> 37 ) & 2047;
-    f->b19 = (serialize_uint32_t) ( rng >> 44 ) & 524287;
-    f->b48 = rng & 0xFFFFFFFFFFFFULL;
-}
-
-static void vary_rt_mixed( RtBenchMixed * f, uint64_t rng )
-{
-    int i;
-    f->sequence = (serialize_uint32_t) ( rng >> 8 ) & 65535;
-    f->ack_sequence = (serialize_int32_t) ( (serialize_uint32_t) ( rng >> 24 ) & 65535 );
-    f->ack_bits = (serialize_uint32_t) ( rng >> 16 );
-    f->session_id = rng;
-    f->client_id = (serialize_uint32_t) ( rng >> 32 );
-    f->nonce = rng ^ 0xA5A5A5A5A5A5A5A5ULL;
-    f->world_time = (serialize_int64_t) ( ( rng >> 12 ) & 0xFFFFFFFFFULL ) - 34359738368LL;
-    f->frame_tick = rng & 0xFFFFFFFFFFFFULL;
-    f->server_time = (serialize_int32_t) ( ( rng >> 20 ) & 0x7FFFFF );
-    for ( i = 0; i < 8; i++ )
-    {
-        RtMixedEntity * e = &f->entities[i];
-        e->entity_id = (serialize_uint32_t) ( ( rng >> i ) & 4095 );
-        e->pos_x = (serialize_int32_t) ( ( rng >> ( i + 4 ) ) & 16383 ) - 8192;
-        e->pos_y = (serialize_int32_t) ( ( rng >> ( i + 12 ) ) & 16383 ) - 8192;
-        e->health = (serialize_int32_t) ( ( rng >> ( i + 20 ) ) & 511 );
-        e->weapon = (serialize_int32_t) ( ( rng >> ( i + 40 ) ) & 15 );
-        e->damage = (serialize_uint32_t) ( ( rng >> ( i + 28 ) ) & 255 );
-        e->moving = ( ( rng >> i ) & 1 ) != 0;
-    }
-    for ( i = 0; i < 80; i++ )
-        f->stats[i].delta = (serialize_int32_t) ( ( rng >> ( i & 31 ) ) & 1023 ) - 512;
-    f->hit.target_id = (serialize_uint32_t) ( ( rng >> 6 ) & 4095 );
-    f->hit.damage = (serialize_int32_t) ( ( rng >> 18 ) & 4095 );
-    f->hit.hit_kind = (serialize_int32_t) ( ( rng >> 30 ) & 7 );
-    f->hit.crit = ( rng & 4 ) != 0;
-    f->loadout[0] = (serialize_uint8_t) ( rng >> 56 );
-    f->player_name[7] = (char) ( 65 + ( ( rng >> 50 ) & 15 ) );
-    f->payload[0] = (serialize_uint8_t) ( rng >> 48 );
-    f->aim_x = (float) ( (serialize_uint32_t) ( rng >> 2 ) & 255 ) * ( 1.0f / 256.0f ) - 0.5f;
-    f->aim_y = (float) ( (serialize_uint32_t) ( rng >> 10 ) & 255 ) * ( 1.0f / 256.0f ) - 0.5f;
-    f->aim_z = (float) ( (serialize_uint32_t) ( rng >> 18 ) & 255 ) * ( 1.0f / 256.0f ) - 0.5f;
-    f->recoil = (float) ( (serialize_uint32_t) rng & 0xFFFF );
-    f->drift = (double) ( (serialize_int64_t) ( ( rng >> 8 ) & 0xFFFFFF ) ) * 0.5;
-    f->wide_key = serialize_uint128_make( rng >> 1, rng );
-    f->flux = serialize_int128_from_int64( (serialize_int64_t) ( rng >> 16 ) );
-    f->ping = (serialize_uint16_t) ( ( rng >> 40 ) & 0x7FFF );
-    f->crc_hint = (serialize_uint32_t) ( ( rng >> 24 ) & 0xFFFFFF );
-    f->extra = (serialize_int32_t) ( ( rng >> 52 ) & 255 );
-}
-
-/* ---- the rt driver: once-helpers carry the single untimed call site per op
-   (§3.2), the noinline loops carry the timed one, and the §1.5 oracle gate
-   runs before any number ---- */
-
-#define BENCH_RT( SUFFIX, TYPE, WRITE, READ, VARY )                                             \
-static int rt_once_write_##SUFFIX( TYPE * msg, uint8_t * buffer )                               \
-{                                                                                               \
-    serialize_write_stream_t stream;                                                            \
-    serialize_write_stream_init( &stream, buffer, BufferSize );                                 \
-    if ( !WRITE( &stream, msg ) )                                                               \
-        return -1;                                                                              \
-    serialize_write_flush( &stream );                                                           \
-    return serialize_write_bytes_processed( &stream );                                          \
-}                                                                                               \
-static int rt_once_read_##SUFFIX( TYPE * out, const uint8_t * buffer, int bytes )               \
-{                                                                                               \
-    serialize_read_stream_t stream;                                                             \
-    serialize_read_stream_init( &stream, buffer, bytes );                                       \
-    return READ( &stream, out );                                                                \
-}                                                                                               \
-BENCH_NOINLINE static int rt_##SUFFIX##_write_loop( TYPE * base, long iters, uint64_t * rng )   \
-{                                                                                               \
-    long i;                                                                                     \
-    for ( i = 0; i < iters; i++ )                                                               \
-    {                                                                                           \
-        serialize_write_stream_t stream;                                                        \
-        *rng = bench_rng( *rng );                                                               \
-        VARY( base, *rng );                                                                     \
-        serialize_write_stream_init( &stream, g_buffer, BufferSize );                           \
-        if ( !WRITE( &stream, base ) )                                                          \
-            return 0;                                                                           \
-        serialize_write_flush( &stream );                                                       \
-        bench_escape( g_buffer );                                                               \
-        g_sink = g_sink + (uint64_t) serialize_write_bytes_processed( &stream );                \
-    }                                                                                           \
-    return 1;                                                                                   \
-}                                                                                               \
-BENCH_NOINLINE static int rt_##SUFFIX##_read_loop( TYPE * out, long iters, int bytes_per_op )   \
-{                                                                                               \
-    long i;                                                                                     \
-    for ( i = 0; i < iters; i++ )                                                               \
-    {                                                                                           \
-        serialize_read_stream_t stream;                                                         \
-        serialize_read_stream_init( &stream, g_variant( i & ( NumVariants - 1 ) ), bytes_per_op ); \
-        if ( !READ( &stream, out ) )                                                            \
-            return 0;                                                                           \
-        bench_escape( out );                                                                    \
-        g_sink = g_sink + 1;                                                                    \
-    }                                                                                           \
-    return 1;                                                                                   \
-}                                                                                               \
-static void bench_rt_##SUFFIX( const char * name, long base_iters, const TYPE * pinned )        \
-{                                                                                               \
-    const long iters = base_iters / IterScale;                                                  \
-    TYPE base = *pinned;                                                                        \
-    TYPE out;                                                                                   \
-    int bytes_per_op;                                                                           \
-    uint64_t rng;                                                                               \
-    double write_rates[MaxNumRuns];                                                             \
-    double read_rates[MaxNumRuns];                                                              \
-    int run, k;                                                                                 \
-                                                                                                \
-    /* oracle 1: the hand-written wire must equal the generated-code golden */                  \
-    bytes_per_op = rt_once_write_##SUFFIX( &base, g_buffer );                                   \
-    if ( bytes_per_op < 0 )                                                                     \
-    {                                                                                           \
-        fail( name, "write of pinned instance failed" );                                        \
-        return;                                                                                 \
-    }                                                                                           \
-    if ( !check_golden( name, g_buffer, bytes_per_op ) )                                        \
-    {                                                                                           \
-        failed = 1;                                                                             \
-        return;                                                                                 \
-    }                                                                                           \
-                                                                                                \
-    /* oracle 2: round-trip write -> read -> re-write -> identical bytes */                     \
-    memset( &out, 0, sizeof( out ) );                                                           \
-    if ( !rt_once_read_##SUFFIX( &out, g_buffer, bytes_per_op ) )                               \
-    {                                                                                           \
-        fail( name, "read of pinned instance failed" );                                         \
-        return;                                                                                 \
-    }                                                                                           \
-    if ( rt_once_write_##SUFFIX( &out, g_twin ) != bytes_per_op ||                              \
-         memcmp( g_buffer, g_twin, (size_t) bytes_per_op ) != 0 )                               \
-    {                                                                                           \
-        fail( name, "round-trip bytes differ" );                                                \
-        return;                                                                                 \
-    }                                                                                           \
-                                                                                                \
-    /* variant buffers (and proof that variation keeps bytes/op constant) */                    \
-    rng = 1;                                                                                    \
-    for ( k = 0; k < NumVariants; k++ )                                                         \
-    {                                                                                           \
-        rng = bench_rng( rng );                                                                 \
-        VARY( &base, rng );                                                                     \
-        if ( rt_once_write_##SUFFIX( &base, g_variant( k ) ) != bytes_per_op )                  \
-        {                                                                                       \
-            fail( name, "variation changed bytes/op — vary must keep structure fields fixed" ); \
-            return;                                                                             \
-        }                                                                                       \
-    }                                                                                           \
-                                                                                                \
-    for ( run = -1; run < g_num_runs; run++ )                                                   \
-    {                                                                                           \
-        double start = time_now();                                                              \
-        double elapsed;                                                                         \
-        if ( !rt_##SUFFIX##_write_loop( &base, iters, &rng ) )                                  \
-        {                                                                                       \
-            fail( name, "write failed in loop" );                                               \
-            return;                                                                             \
-        }                                                                                       \
-        elapsed = time_now() - start;                                                           \
-        if ( run >= 0 )                                                                         \
-            write_rates[run] = (double) iters / elapsed;                                        \
-    }                                                                                           \
-                                                                                                \
-    for ( run = -1; run < g_num_runs; run++ )                                                   \
-    {                                                                                           \
-        double start = time_now();                                                              \
-        double elapsed;                                                                         \
-        if ( !rt_##SUFFIX##_read_loop( &out, iters, bytes_per_op ) )                            \
-        {                                                                                       \
-            fail( name, "read failed in loop" );                                                \
-            return;                                                                             \
-        }                                                                                       \
-        elapsed = time_now() - start;                                                           \
-        if ( run >= 0 )                                                                         \
-            read_rates[run] = (double) iters / elapsed;                                         \
-    }                                                                                           \
-                                                                                                \
-    report( name, "write", iters, bytes_per_op, run_stats( write_rates, g_num_runs ), "rt" );   \
-    report( name, "read", iters, bytes_per_op, run_stats( read_rates, g_num_runs ), "rt" );     \
-}
-
-BENCH_RT( bench_packet, RtBenchPacket, rt_write_bench_packet, rt_read_bench_packet, vary_rt_packet )
-BENCH_RT( bench_ints, RtBenchInts, rt_write_bench_ints, rt_read_bench_ints, vary_rt_ints )
-BENCH_RT( bench_bits, RtBenchBits, rt_write_bench_bits, rt_read_bench_bits, vary_rt_bits )
-BENCH_RT( bench_mixed, RtBenchMixed, rt_write_bench_mixed, rt_read_bench_mixed, vary_rt_mixed )
 
 /* ------------------------------------------------------------------------------------------
    family bits (BENCH-STANDARD.md §1.4): the raw bit surface — serialize.c
@@ -1764,12 +1120,10 @@ int main( int argc, char ** argv )
         bench_message_real_packet( "real_packet", "real_packet", 8000000L, &real_packet );
     }
 
-    /* family gen over the Bench corpus (issue #177): the generated twins of
-       the rt rows below — same shapes, same goldens, same pins, same vary
+    /* family gen over the Bench corpus (issue #177): the four Bench.schema
+       shapes through the generated code — same goldens, same pins, same vary
        mappings, same iteration counts (fixed and identical across all five
-       runners, §2.1); only the subject differs, and the family column says
-       so. --quick runs the gen bench_mixed (the schema subject) beside the
-       rt row (hand-written usage). */
+       runners, §2.1). --quick runs the gen bench_mixed. */
     {
         BenchPacket gen_packet;
         BenchInts gen_ints;
@@ -1799,89 +1153,8 @@ int main( int argc, char ** argv )
             bench_message_gen_bench_ints( "bench_ints", "bench_ints", 40000000L, &gen_ints );
             bench_message_gen_bench_bits( "bench_bits", "bench_bits", 48000000L, &gen_bits );
         }
-        /* --quick runs exactly this gen leg beside the rt one below */
+        /* --quick runs exactly this gen leg */
         bench_datadriven_gen_bench_mixed( "bench_mixed", "bench_mixed", 4000000L );
-    }
-
-    /* family rt (§1.3/§1.5): the runtime API by hand, oracle-gated against
-       the goldens the generated code pinned. Iteration counts are fixed and
-       identical across all five runners (§2.1; sized in the C++ reference). */
-    {
-        RtBenchPacket rt_packet;
-        RtBenchInts rt_ints;
-        RtBenchBits rt_bits;
-        RtBenchMixed rt_mixed;
-
-        memset( &rt_packet, 0, sizeof( rt_packet ) );
-        rt_packet.a = -37; rt_packet.b = 12345; rt_packet.c = 987654;
-        rt_packet.bits7 = 97; rt_packet.bits13 = 5000; rt_packet.bits23 = 1234567;
-        rt_packet.flag = 1;
-        rt_packet.x = 1.5f; rt_packet.y = -3.25f; rt_packet.z = 100.125f;
-        rt_packet.big = 0x123456789ABCDEF0ULL;
-        for ( i = 0; i < 17; i++ )
-            rt_packet.blob[i] = (serialize_uint8_t) ( i * 31 );
-
-        memset( &rt_ints, 0, sizeof( rt_ints ) );
-        rt_ints.f0 = -37; rt_ints.f1 = 12345; rt_ints.f2 = 987654; rt_ints.f3 = 2; rt_ints.f4 = -15;
-        rt_ints.f5 = 777; rt_ints.f6 = -2048; rt_ints.f7 = 200; rt_ints.f8 = -543210; rt_ints.f9 = 99;
-
-        memset( &rt_bits, 0, sizeof( rt_bits ) );
-        rt_bits.b7 = 97; rt_bits.b13 = 5000; rt_bits.b23 = 1234567; rt_bits.b3 = 5;
-        rt_bits.b32 = 0xDEADBEEFu; rt_bits.b11 = 1024; rt_bits.b19 = 333333;
-        rt_bits.b48 = 0xFEDCBA987654ULL;
-
-        memset( &rt_mixed, 0, sizeof( rt_mixed ) );
-        rt_mixed.magic = 0xC0DE;
-        rt_mixed.sequence = 52428; rt_mixed.ack_sequence = 12345; rt_mixed.ack_bits = 0xA5A5A5A5u;
-        rt_mixed.session_id = 0x123456789ABCDEF0ULL; rt_mixed.client_id = 0xDEADBEEFu;
-        rt_mixed.nonce = 0xFEDCBA9876543210ULL; rt_mixed.world_time = -987654321000LL;
-        rt_mixed.frame_tick = 0x123456789ABCULL; rt_mixed.server_time = 12345678;
-        rt_mixed.entities_count = 8;
-        for ( i = 0; i < 8; i++ )
-        {
-            rt_mixed.entities[i].entity_id = (serialize_uint32_t) ( 2049 + i * 17 );
-            rt_mixed.entities[i].pos_x = -16383 + i * 4096;
-            rt_mixed.entities[i].pos_y = 16383 - i * 4096;
-            rt_mixed.entities[i].pos_z = -1 + i * 2048;
-            rt_mixed.entities[i].yaw = (serialize_uint32_t) ( 511 - i * 64 );
-            rt_mixed.entities[i].pitch = (serialize_uint32_t) ( i * 73 );
-            rt_mixed.entities[i].vel_x = -2048 + i * 512;
-            rt_mixed.entities[i].vel_y = 2047 - i * 512;
-            rt_mixed.entities[i].vel_z = -1024 + i * 256;
-            rt_mixed.entities[i].health = 1000 - i * 100;
-            rt_mixed.entities[i].weapon = 1 + i;
-            rt_mixed.entities[i].damage = (serialize_uint32_t) ( 0x5A + i );
-            rt_mixed.entities[i].moving = ( i % 2 ) == 0;
-            rt_mixed.entities[i].firing = ( i % 3 ) == 0;
-        }
-        rt_mixed.stats_count = 80;
-        for ( i = 0; i < 80; i++ )
-        {
-            rt_mixed.stats[i].stat_id = (serialize_uint32_t) ( ( i * 3 ) % 256 );
-            rt_mixed.stats[i].delta = -512 + ( i * 13 ) % 1024;
-        }
-        rt_mixed.event_type = 1;   /* Hit */
-        rt_mixed.hit.target_id = 4095; rt_mixed.hit.damage = 4095;
-        rt_mixed.hit.hit_kind = 7; rt_mixed.hit.crit = 1;
-        rt_mixed.loadout[0] = 0x11; rt_mixed.loadout[1] = 0x22;
-        rt_mixed.loadout[2] = 0x33; rt_mixed.loadout[3] = 0x44;
-        memcpy( rt_mixed.player_name, "Rowan_01", 8 ); rt_mixed.player_name_length = 8;
-        memcpy( rt_mixed.payload, "\xDE\xAD\xBE\xEF\x01\x02\x03\x04", 8 ); rt_mixed.payload_length = 8;
-        rt_mixed.aim_x = 0.5f; rt_mixed.aim_y = -0.25f; rt_mixed.aim_z = 0.75f;
-        rt_mixed.recoil = 1.5f; rt_mixed.drift = -3.25;
-        rt_mixed.wide_key = serialize_uint128_make( 0x0123456789ABCDEFULL, 0xFEDCBA9876543210ULL );
-        rt_mixed.flux = serialize_int128_make( 0x800000000ULL, 7ULL );   /* 2^99 + 7 */
-        rt_mixed.ping = 12345; rt_mixed.crc_hint = 0xABCDEFu;
-        rt_mixed.has_extra = 1; rt_mixed.extra = 200;
-
-        if ( !g_quick )
-        {
-            bench_rt_bench_packet( "bench_packet", 32000000L, &rt_packet );
-            bench_rt_bench_ints( "bench_ints", 40000000L, &rt_ints );
-            bench_rt_bench_bits( "bench_bits", 48000000L, &rt_bits );
-        }
-        /* --quick runs exactly this leg: bench_mixed, golden gate intact */
-        bench_rt_bench_mixed( "bench_mixed", 4000000L, &rt_mixed );
     }
 
     /* family bits (§1.4): the one bitpacker workload in the estate */
