+++ bench/dart/main.dart
@@ -13,9 +13,8 @@
 // WHAT THE ROWS MEASURE: the generated codec — schema's Dart backend emits
 // self-contained monomorphic write/read functions with the bitpacker
 // inlined and zero runtime dependencies, so the generated code IS the Dart
-// serialize path. The rows carry family=gen: a ratio against another
-// language's family=rt row (the serialize runtime API called by hand) is a
-// subject difference, not a language difference, and the tools refuse it.
+// serialize path. The rows carry family=gen — the estate's one benchmark
+// subject (schema#196).
 // Peak-style numbers from this runner's earlier serialize-family form
 // (tight per-shape loops, best-of-five) are NOT comparable to these rows —
 // different measurement contract, and the statistic alone moves the number.
