+++ bench/js/main.mjs
@@ -7,9 +7,7 @@
 // mismatch REFUSES to bench), same warmup + 7 measured runs +
 // median/min/max/spread, same CSV row format with lang=js. See
 // bench/README.md for the runner contract. One module, like bench/c's one
-// translation unit: the harness helpers are module-scoped, and splitting the
-// rt family into a second module would force them through an export surface
-// the other runners do not have.
+// translation unit: the harness helpers are module-scoped.
 //
 // TWO GENERATED CODECS ride this runner (BENCH-STANDARD.md §5.1 codec
 // column, 2026-08-18): the gen-family rows measure the FLAT tier
@@ -42,9 +40,6 @@
 //   - no alloc note: Node exposes no per-thread allocation counter the Go
 //     and C# notes read; the reuse discipline here is structural (persistent
 //     holders, stream reset, pre-bounded variant views)
-//   - family rt holds fields as persistent { value } holder objects — the
-//     serialize.js caller idiom, from the runtime's own bench — so the
-//     timed loops allocate nothing of their own
 //
 // Both runtime modes run: NODE_ENV forks serialize.js at module load
 // (src/mode.js). The standard leg is NODE_ENV=production — the caller-trust
@@ -87,9 +82,6 @@ import * as realworldFlat from "../../generated/bench/js/realworld/RealWorldFlat
 // the four bench/corpus/Bench.schema shapes as GENERATED code: the flat
 // tier is THE js path for these shapes exactly as for the corpus shapes
 // above, cross-validated against the runtime-call tier by the same oracle.
-// The family rt rows for the same shapes (the serialize.js runtime API
-// called by hand, below) ride beside them as library-context data — the
-// same bench names, distinguished by family and codec.
 import * as bench from "../../generated/bench/js/Bench.js";
 import * as benchFlat from "../../generated/bench/js/BenchFlat.js";
 
@@ -139,7 +131,7 @@ const NumVariants = 64; // read-path variant buffers
 // levels), inline unknown — and it stays unknown: a JIT leg has no §4.1
 // AOT artifact to disassemble, so the verdict pass has no js branch and js
 // rows never ratio against a row whose inline column is filled.
-// family is per ROW (gen | rt | bits — §5.1).
+// family is per ROW (gen | bits — §5.1).
 const CsvSuffix = `esm,${PRODUCTION ? "contract" : "always"},default,unknown`;
 
 const gCsvRows = []; // { row, family }
@@ -344,50 +336,6 @@ function sinkOfRealPacket(d) {
     d.F096Bits + d.F097Bits;
 }
 
-// family rt: the holder-object shapes (see makeRt*)
-function sinkOfRtPacket(f) {
-  return f.a.value + f.b.value + f.c.value + f.bits7.value + f.bits13.value +
-    f.bits23.value + boolBit(f.flag.value) + f.x.value + f.y.value + f.z.value +
-    bigBit(f.big.value) + sumBytes(f.blob, 17);
-}
-
-function sinkOfRtInts(f) {
-  return f.f0.value + f.f1.value + f.f2.value + f.f3.value + f.f4.value +
-    f.f5.value + f.f6.value + f.f7.value + f.f8.value + f.f9.value;
-}
-
-function sinkOfRtBits(f) {
-  return f.b7.value + f.b13.value + f.b23.value + f.b3.value + f.b32.value +
-    f.b11.value + f.b19.value + bigBit(f.b48.value);
-}
-
-// §2.7 full-struct observation over the hand-written shape — the gen sink's twin
-function sinkOfRtMixed(f) {
-  let s = f.magic.value + f.sequence.value + f.ackSequence.value + f.ackBits.value +
-    bigBit(f.sessionId.value) + f.clientId.value + bigBit(f.nonce.value) +
-    bigBit(f.worldTime.value) + bigBit(f.frameTick.value) + f.serverTime.value +
-    f.entitiesCount.value + f.statsCount.value + f.eventType.value +
-    f.playerNameLength.value + f.payloadLength.value +
-    f.aimX.value + f.aimY.value + f.aimZ.value + f.recoil.value + f.drift.value +
-    bigBit(f.wideKey.value) + bigBit(f.flux.value) + f.ping.value +
-    f.reservedBits.value + f.crcHint.value + boolBit(f.hasExtra.value) +
-    f.extra.value + f.idleTicks.value;
-  for (let i = 0; i < f.entitiesCount.value; i++) {
-    const e = f.entities[i];
-    s += e.entityId.value + e.posX.value + e.posY.value + e.posZ.value +
-      e.yaw.value + e.pitch.value + e.velX.value + e.velY.value + e.velZ.value +
-      e.health.value + e.weapon.value + e.damage.value +
-      boolBit(e.moving.value) + boolBit(e.firing.value);
-  }
-  for (let i = 0; i < f.statsCount.value; i++) {
-    s += f.stats[i].statId.value + f.stats[i].delta.value;
-  }
-  s += f.hitTargetId.value + f.hitDamage.value + f.hitKind.value + boolBit(f.hitCrit.value);
-  s += sumBytes(f.loadout, 4);
-  s += sumBytes(f.playerName, f.playerNameLength.value);
-  s += sumBytes(f.payload, f.payloadLength.value);
-  return s;
-}
 let gCsv = false;
 let gWireDir = "../../testdata/wire";
 let gVariantDir = "../../bench/corpus/variants";
@@ -1291,765 +1239,6 @@ function varyRealPacket(m) {
 
 
 // --------------------------------------------------------------------------
-// family rt (BENCH-STANDARD.md §1.3, §1.5): the serialize.js runtime API
-// called BY HAND — the four Bench.schema shapes as hand-written packets over
-// the plain WriteStream/ReadStream serialize* surface, the way a game would
-// write them. Fields are persistent { value } holders (the serialize.js
-// caller idiom, from the runtime's own bench), created once and mutated in
-// place, so the loops allocate nothing of their own. The §1.5 oracle gate
-// byte-compares the hand-written wire against the goldens the GENERATED
-// code pinned (testdata/wire/bench_*.bin) and round-trips before any
-// number. Per §3.2 every benched op has EXACTLY two call sites: its untimed
-// once-helper and its timed loop — per-shape functions, as in the Go and C#
-// runners.
-// --------------------------------------------------------------------------
-
-function makeRtPacket() {
-  return {
-    a: { value: 0 },
-    b: { value: 0 },
-    c: { value: 0 },
-    bits7: { value: 0 },
-    bits13: { value: 0 },
-    bits23: { value: 0 },
-    flag: { value: false },
-    x: { value: 0 },
-    y: { value: 0 },
-    z: { value: 0 },
-    big: { value: 0n },
-    blob: new Uint8Array(17),
-  };
-}
-
-function writeRtPacket(s, p) {
-  return (
-    s.serializeInt(p.a, -100, 100) &&
-    s.serializeInt(p.b, 0, 65535) &&
-    s.serializeInt(p.c, -1000000, 1000000) &&
-    s.serializeBits(p.bits7, 7) &&
-    s.serializeBits(p.bits13, 13) &&
-    s.serializeBits(p.bits23, 23) &&
-    s.serializeBool(p.flag) &&
-    s.serializeFloat(p.x) &&
-    s.serializeFloat(p.y) &&
-    s.serializeFloat(p.z) &&
-    s.serializeUint64(p.big) &&
-    s.serializeBytes(p.blob) // aligns internally — the schema says `align` out loud
-  );
-}
-
-function readRtPacket(s, p) {
-  return (
-    s.serializeInt(p.a, -100, 100) &&
-    s.serializeInt(p.b, 0, 65535) &&
-    s.serializeInt(p.c, -1000000, 1000000) &&
-    s.serializeBits(p.bits7, 7) &&
-    s.serializeBits(p.bits13, 13) &&
-    s.serializeBits(p.bits23, 23) &&
-    s.serializeBool(p.flag) &&
-    s.serializeFloat(p.x) &&
-    s.serializeFloat(p.y) &&
-    s.serializeFloat(p.z) &&
-    s.serializeUint64(p.big) &&
-    s.serializeBytes(p.blob)
-  );
-}
-
-function makeRtInts() {
-  return {
-    f0: { value: 0 },
-    f1: { value: 0 },
-    f2: { value: 0 },
-    f3: { value: 0 },
-    f4: { value: 0 },
-    f5: { value: 0 },
-    f6: { value: 0 },
-    f7: { value: 0 },
-    f8: { value: 0 },
-    f9: { value: 0 },
-  };
-}
-
-function writeRtInts(s, f) {
-  return (
-    s.serializeInt(f.f0, -100, 100) &&
-    s.serializeInt(f.f1, 0, 65535) &&
-    s.serializeInt(f.f2, -1000000, 1000000) &&
-    s.serializeInt(f.f3, 0, 3) &&
-    s.serializeInt(f.f4, -15, 15) &&
-    s.serializeInt(f.f5, 0, 1000) &&
-    s.serializeInt(f.f6, -2048, 2047) &&
-    s.serializeInt(f.f7, 0, 255) &&
-    s.serializeInt(f.f8, -600000, 600000) &&
-    s.serializeInt(f.f9, 0, 100)
-  );
-}
-
-function readRtInts(s, f) {
-  return (
-    s.serializeInt(f.f0, -100, 100) &&
-    s.serializeInt(f.f1, 0, 65535) &&
-    s.serializeInt(f.f2, -1000000, 1000000) &&
-    s.serializeInt(f.f3, 0, 3) &&
-    s.serializeInt(f.f4, -15, 15) &&
-    s.serializeInt(f.f5, 0, 1000) &&
-    s.serializeInt(f.f6, -2048, 2047) &&
-    s.serializeInt(f.f7, 0, 255) &&
-    s.serializeInt(f.f8, -600000, 600000) &&
-    s.serializeInt(f.f9, 0, 100)
-  );
-}
-
-function makeRtBits() {
-  return {
-    b7: { value: 0 },
-    b13: { value: 0 },
-    b23: { value: 0 },
-    b3: { value: 0 },
-    b32: { value: 0 },
-    b11: { value: 0 },
-    b19: { value: 0 },
-    b48: { value: 0n },
-  };
-}
-
-function writeRtBits(s, f) {
-  return (
-    s.serializeBits(f.b7, 7) &&
-    s.serializeBits(f.b13, 13) &&
-    s.serializeBits(f.b23, 23) &&
-    s.serializeBits(f.b3, 3) &&
-    s.serializeBits(f.b32, 32) &&
-    s.serializeBits(f.b11, 11) &&
-    s.serializeBits(f.b19, 19) &&
-    s.serializeBits64(f.b48, 48)
-  );
-}
-
-function readRtBits(s, f) {
-  return (
-    s.serializeBits(f.b7, 7) &&
-    s.serializeBits(f.b13, 13) &&
-    s.serializeBits(f.b23, 23) &&
-    s.serializeBits(f.b3, 3) &&
-    s.serializeBits(f.b32, 32) &&
-    s.serializeBits(f.b11, 11) &&
-    s.serializeBits(f.b19, 19) &&
-    s.serializeBits64(f.b48, 48)
-  );
-}
-
-// BenchMixed by hand (issue #184): every serialize runtime operation the
-// schema language expresses, in the order the generated code emits them.
-// The §1.5 oracle gate byte-compares this against the generated golden.
-const RT_FLUX_MIN = -(1n << 100n);
-const RT_FLUX_MAX = 1n << 100n;
-
-function makeRtMixedEntity() {
-  return {
-    entityId: { value: 0 },
-    posX: { value: 0 },
-    posY: { value: 0 },
-    posZ: { value: 0 },
-    yaw: { value: 0 },
-    pitch: { value: 0 },
-    velX: { value: 0 },
-    velY: { value: 0 },
-    velZ: { value: 0 },
-    health: { value: 0 },
-    weapon: { value: 0 },
-    damage: { value: 0 },
-    moving: { value: false },
-    firing: { value: false },
-  };
-}
-
-function makeRtMixed() {
-  const playerName = new Uint8Array(15);
-  const payload = new Uint8Array(16);
-  return {
-    magic: { value: 0xc0de },
-    sequence: { value: 0 },
-    ackSequence: { value: 0 },
-    ackBits: { value: 0 },
-    sessionId: { value: 0n },
-    clientId: { value: 0 },
-    nonce: { value: 0n },
-    worldTime: { value: 0n },
-    frameTick: { value: 0n },
-    serverTime: { value: 0 },
-    entitiesCount: { value: 0 },
-    entities: Array.from({ length: 8 }, makeRtMixedEntity),
-    statsCount: { value: 0 },
-    stats: Array.from({ length: 80 }, () => ({ statId: { value: 0 }, delta: { value: 0 } })),
-    eventType: { value: 0 },
-    hitTargetId: { value: 0 },
-    hitDamage: { value: 0 },
-    hitKind: { value: 0 },
-    hitCrit: { value: false },
-    chatChannel: { value: 0 },
-    chatSpeaker: { value: 0 },
-    pickupItemId: { value: 0 },
-    pickupAmount: { value: 0 },
-    loadout: new Uint8Array(4),
-    playerNameLength: { value: 0 },
-    playerName,
-    // the used lengths are STRUCTURE, so the subarray views are built ONCE
-    // here rather than per call (serializeBytes takes a Uint8Array)
-    playerNameUsed: playerName.subarray(0, 8),
-    payloadLength: { value: 0 },
-    payload,
-    payloadUsed: payload.subarray(0, 8),
-    aimX: { value: 0 },
-    aimY: { value: 0 },
-    aimZ: { value: 0 },
-    recoil: { value: 0 },
-    drift: { value: 0 },
-    wideKey: { value: 0n },
-    flux: { value: 0n },
-    ping: { value: 0 },
-    reservedBits: { value: 0 },
-    crcHint: { value: 0 },
-    hasExtra: { value: true },
-    extra: { value: 0 },
-    idleTicks: { value: 0 },
-  };
-}
-
-function serializeRtMixed(s, f) {
-  if (!s.serializeBits(f.magic, 16)) return false;
-  if (!s.serializeBits(f.sequence, 16)) return false;
-  if (!s.serializeInt(f.ackSequence, 0, 65535)) return false;
-  if (!s.serializeBits(f.ackBits, 32)) return false;
-  if (!s.serializeUint64(f.sessionId)) return false;
-  if (!s.serializeUint32(f.clientId)) return false;
-  // the full-unsigned ranged path is width-computed raw bits
-  if (!s.serializeBits64(f.nonce, 64)) return false;
-  if (!s.serializeInt64(f.worldTime, -1000000000000n, 1000000000000n)) return false;
-  if (!s.serializeBits64(f.frameTick, 48)) return false;
-  if (!s.serializeFixed(f.serverTime, 24, 8, 0, 65535)) return false;
-
-  if (!s.serializeInt(f.entitiesCount, 1, 8)) return false;
-  for (let i = 0; i < f.entitiesCount.value; i++) {
-    const e = f.entities[i];
-    if (!s.serializeBits(e.entityId, 12)) return false;
-    if (!s.serializeInt(e.posX, -16383, 16383)) return false;
-    if (!s.serializeInt(e.posY, -16383, 16383)) return false;
-    if (!s.serializeInt(e.posZ, -16383, 16383)) return false;
-    if (!s.serializeBits(e.yaw, 9)) return false;
-    if (!s.serializeBits(e.pitch, 9)) return false;
-    if (!s.serializeInt(e.velX, -2048, 2047)) return false;
-    if (!s.serializeInt(e.velY, -2048, 2047)) return false;
-    if (!s.serializeInt(e.velZ, -2048, 2047)) return false;
-    if (!s.serializeInt(e.health, 0, 1000)) return false;
-    if (!s.serializeInt(e.weapon, 0, 15)) return false;
-    if (!s.serializeBits(e.damage, 8)) return false;
-    if (!s.serializeBool(e.moving)) return false;
-    if (!s.serializeBool(e.firing)) return false;
-  }
-
-  if (!s.serializeInt(f.statsCount, 0, 80)) return false;
-  for (let i = 0; i < f.statsCount.value; i++) {
-    if (!s.serializeBits(f.stats[i].statId, 8)) return false;
-    if (!s.serializeInt(f.stats[i].delta, -512, 511)) return false;
-  }
-
-  if (!s.serializeInt(f.eventType, 0, 3)) return false;
-  if (f.eventType.value === 1) {
-    if (!s.serializeBits(f.hitTargetId, 12)) return false;
-    if (!s.serializeInt(f.hitDamage, 0, 4095)) return false;
-    if (!s.serializeInt(f.hitKind, 0, 7)) return false;
-    if (!s.serializeBool(f.hitCrit)) return false;
-  } else if (f.eventType.value === 2) {
-    if (!s.serializeInt(f.chatChannel, 0, 3)) return false;
-    if (!s.serializeBits(f.chatSpeaker, 12)) return false;
-  } else if (f.eventType.value === 3) {
-    if (!s.serializeBits(f.pickupItemId, 10)) return false;
-    if (!s.serializeInt(f.pickupAmount, 0, 255)) return false;
-  }
-
-  for (let i = 0; i < 4; i++) {
-    RT_BYTE.value = f.loadout[i];
-    if (!s.serializeUint8(RT_BYTE)) return false;
-    f.loadout[i] = RT_BYTE.value;
-  }
-
-  // string(15) and bytes(16) ride as their §4.3 decomposition in every rt
-  // leg — see bench/cpp/bench_main.cpp for the reasoning
-  if (!s.serializeInt(f.playerNameLength, 0, 15)) return false;
-  if (!s.serializeBytes(f.playerNameUsed)) return false;
-  if (!s.serializeInt(f.payloadLength, 0, 16)) return false;
-  if (!s.serializeBytes(f.payloadUsed)) return false;
-
-  if (!s.serializeCompressedFloat(f.aimX, -1.0, 1.0, 0.01)) return false;
-  if (!s.serializeCompressedFloat(f.aimY, -1.0, 1.0, 0.01)) return false;
-  if (!s.serializeCompressedFloat(f.aimZ, -1.0, 1.0, 0.01)) return false;
-  if (!s.serializeFloat(f.recoil)) return false;
-  if (!s.serializeDouble(f.drift)) return false;
-  if (!s.serializeUint128(f.wideKey)) return false;
-  if (!s.serializeInt128(f.flux, RT_FLUX_MIN, RT_FLUX_MAX)) return false;
-  if (!s.serializeFixed(f.ping, 8, 8, 0, 250)) return false;
-
-  if (!s.serializeBits(f.reservedBits, 4)) return false;
-  if (!s.serializeAlign()) return false;
-  if (!s.serializeBits(f.crcHint, 24)) return false;
-  if (!s.serializeBool(f.hasExtra)) return false;
-  if (f.hasExtra.value) {
-    if (!s.serializeInt(f.extra, 0, 255)) return false;
-  } else {
-    if (!s.serializeInt(f.idleTicks, 0, 15)) return false;
-  }
-  return true;
-}
-
-// serializeUint8 takes a ref, and the loadout lives in a Uint8Array; one
-// shared scratch ref keeps the loop allocation-free
-const RT_BYTE = { value: 0 };
-
-function writeRtMixed(s, f) {
-  return serializeRtMixed(s, f);
-}
-
-function readRtMixed(s, f) {
-  if (!serializeRtMixed(s, f)) return false;
-  // const(0xC0DE, 16) and reserved(4) are contract fields the runtime API
-  // does not check for you; the generated reader refuses the same bytes
-  return f.magic.value === 0xc0de && f.reservedBits.value === 0;
-}
-
-// ---- pinned instances: test/bench/main.cpp (the golden producer), verbatim ----
-
-function pinRtPacket() {
-  const p = makeRtPacket();
-  p.a.value = -37;
-  p.b.value = 12345;
-  p.c.value = 987654;
-  p.bits7.value = 97;
-  p.bits13.value = 5000;
-  p.bits23.value = 1234567;
-  p.flag.value = true;
-  p.x.value = 1.5;
-  p.y.value = -3.25;
-  p.z.value = 100.125;
-  p.big.value = 0x123456789abcdef0n;
-  for (let i = 0; i < 17; i++) {
-    p.blob[i] = (i * 31) & 0xff;
-  }
-  return p;
-}
-
-function pinRtInts() {
-  const f = makeRtInts();
-  f.f0.value = -37;
-  f.f1.value = 12345;
-  f.f2.value = 987654;
-  f.f3.value = 2;
-  f.f4.value = -15;
-  f.f5.value = 777;
-  f.f6.value = -2048;
-  f.f7.value = 200;
-  f.f8.value = -543210;
-  f.f9.value = 99;
-  return f;
-}
-
-function pinRtBits() {
-  const f = makeRtBits();
-  f.b7.value = 97;
-  f.b13.value = 5000;
-  f.b23.value = 1234567;
-  f.b3.value = 5;
-  f.b32.value = 0xdeadbeef;
-  f.b11.value = 1024;
-  f.b19.value = 333333;
-  f.b48.value = 0xfedcba987654n;
-  return f;
-}
-
-function pinRtMixed() {
-  const f = makeRtMixed();
-  f.magic.value = 0xc0de;
-  f.sequence.value = 52428;
-  f.ackSequence.value = 12345;
-  f.ackBits.value = 0xa5a5a5a5;
-  f.sessionId.value = 0x123456789abcdef0n;
-  f.clientId.value = 0xdeadbeef;
-  f.nonce.value = 0xfedcba9876543210n;
-  f.worldTime.value = -987654321000n;
-  f.frameTick.value = 0x123456789abcn;
-  f.serverTime.value = 12345678;
-  f.entitiesCount.value = 8;
-  for (let i = 0; i < 8; i++) {
-    const e = f.entities[i];
-    e.entityId.value = 2049 + i * 17;
-    e.posX.value = -16383 + i * 4096;
-    e.posY.value = 16383 - i * 4096;
-    e.posZ.value = -1 + i * 2048;
-    e.yaw.value = 511 - i * 64;
-    e.pitch.value = i * 73;
-    e.velX.value = -2048 + i * 512;
-    e.velY.value = 2047 - i * 512;
-    e.velZ.value = -1024 + i * 256;
-    e.health.value = 1000 - i * 100;
-    e.weapon.value = 1 + i;
-    e.damage.value = 0x5a + i;
-    e.moving.value = i % 2 === 0;
-    e.firing.value = i % 3 === 0;
-  }
-  f.statsCount.value = 80;
-  for (let i = 0; i < 80; i++) {
-    f.stats[i].statId.value = (i * 3) % 256;
-    f.stats[i].delta.value = -512 + ((i * 13) % 1024);
-  }
-  f.eventType.value = 1; // Hit
-  f.hitTargetId.value = 4095;
-  f.hitDamage.value = 4095;
-  f.hitKind.value = 7;
-  f.hitCrit.value = true;
-  f.loadout.set([0x11, 0x22, 0x33, 0x44]);
-  f.playerName.set(PLAYER_NAME_PIN);
-  f.playerNameLength.value = 8;
-  f.payload.set(PAYLOAD_PIN);
-  f.payloadLength.value = 8;
-  f.aimX.value = 0.5;
-  f.aimY.value = -0.25;
-  f.aimZ.value = 0.75;
-  f.recoil.value = 1.5;
-  f.drift.value = -3.25;
-  f.wideKey.value = (0x0123456789abcdefn << 64n) | 0xfedcba9876543210n;
-  f.flux.value = (1n << 99n) + 7n;
-  f.ping.value = 12345;
-  f.crcHint.value = 0xabcdef;
-  f.hasExtra.value = true;
-  f.extra.value = 200;
-  return f;
-}
-
-// ---- vary functions: bench/cpp/bench_main.cpp's rt mappings exactly ----
-
-function varyRtPacket(p) {
-  p.a.value = (shr64(8) & 63) - 32;
-  p.b.value = shr64(16) & 65535;
-  p.c.value = (shr64(24) & 0xfffff) - 500000;
-  p.bits7.value = rng.lo & 127;
-  p.bits13.value = shr64(3) & 8191;
-  p.bits23.value = shr64(5) & 8388607;
-  p.flag.value = (rng.lo & 1) !== 0;
-  p.x.value = rng.lo & 0xffff;
-  p.big.value = rngBig();
-  p.blob[0] = shr64(32) & 0xff;
-}
-
-function varyRtInts(f) {
-  f.f0.value = (shr64(8) & 63) - 32;
-  f.f1.value = shr64(16) & 65535;
-  f.f2.value = (shr64(24) & 0xfffff) - 500000;
-  f.f3.value = shr64(2) & 3;
-  f.f4.value = (shr64(11) & 15) - 8;
-  f.f5.value = shr64(22) & 511;
-  f.f6.value = (shr64(33) & 2047) - 1024;
-  f.f7.value = shr64(40) & 255;
-  f.f8.value = (shr64(30) & 0xfffff) - 500000;
-  f.f9.value = shr64(57) & 63;
-}
-
-function varyRtBits(f) {
-  f.b7.value = rng.lo & 127;
-  f.b13.value = shr64(3) & 8191;
-  f.b23.value = shr64(5) & 8388607;
-  f.b3.value = shr64(29) & 7;
-  f.b32.value = shr64(16);
-  f.b11.value = shr64(37) & 2047;
-  f.b19.value = shr64(44) & 524287;
-  f.b48.value = rngBig() & 0xffffffffffffn;
-}
-
-function varyRtMixed(f) {
-  const big = rngBig();
-  f.sequence.value = shr64(8) & 65535;
-  f.ackSequence.value = shr64(24) & 65535;
-  f.ackBits.value = shr64(16);
-  f.sessionId.value = big;
-  f.clientId.value = shr64(32);
-  f.nonce.value = big;
-  f.worldTime.value = BIG_WORLD[shr64(12) & 255];
-  f.frameTick.value = big & 0xffffffffffffn;
-  f.serverTime.value = shr64(20) & 0x7fffff;
-  for (let i = 0; i < 8; i++) {
-    const e = f.entities[i];
-    e.entityId.value = shr64(i) & 4095;
-    e.posX.value = (shr64(i + 4) & 16383) - 8192;
-    e.posY.value = (shr64(i + 12) & 16383) - 8192;
-    e.health.value = shr64(i + 20) & 511;
-    e.weapon.value = shr64(i + 40) & 15;
-    e.damage.value = shr64(i + 28) & 255;
-    e.moving.value = (shr64(i) & 1) !== 0;
-  }
-  for (let i = 0; i < 80; i++) {
-    f.stats[i].delta.value = (shr64(i & 31) & 1023) - 512;
-  }
-  f.hitTargetId.value = shr64(6) & 4095;
-  f.hitDamage.value = shr64(18) & 4095;
-  f.hitKind.value = shr64(30) & 7;
-  f.hitCrit.value = (rng.lo & 4) !== 0;
-  f.loadout[0] = shr64(56) & 255;
-  f.playerName[7] = 65 + (shr64(50) & 15);
-  f.payload[0] = shr64(48) & 255;
-  f.aimX.value = (shr64(2) & 255) * (1 / 256) - 0.5;
-  f.aimY.value = (shr64(10) & 255) * (1 / 256) - 0.5;
-  f.aimZ.value = (shr64(18) & 255) * (1 / 256) - 0.5;
-  f.recoil.value = rng.lo & 0xffff;
-  f.drift.value = (shr64(8) & 0xffffff) * 0.5;
-  f.wideKey.value = big;
-  f.flux.value = BIG_FLUX[shr64(16) & 255];
-  f.ping.value = shr64(40) & 0x7fff;
-  f.crcHint.value = shr64(24) & 0xffffff;
-  f.extra.value = shr64(52) & 255;
-}
-
-// ---- the single untimed call sites (§3.2), one pair per shape ----
-
-function rtOnceWritePacket(p, buf) {
-  const ws = new WriteStream(buf);
-  if (!writeRtPacket(ws, p)) {
-    return -1;
-  }
-  ws.flush();
-  return ws.bytesProcessed();
-}
-
-function rtOnceReadPacket(p, view) {
-  const rs = new ReadStream(view);
-  return readRtPacket(rs, p);
-}
-
-function rtOnceWriteInts(f, buf) {
-  const ws = new WriteStream(buf);
-  if (!writeRtInts(ws, f)) {
-    return -1;
-  }
-  ws.flush();
-  return ws.bytesProcessed();
-}
-
-function rtOnceReadInts(f, view) {
-  const rs = new ReadStream(view);
-  return readRtInts(rs, f);
-}
-
-function rtOnceWriteBits(f, buf) {
-  const ws = new WriteStream(buf);
-  if (!writeRtBits(ws, f)) {
-    return -1;
-  }
-  ws.flush();
-  return ws.bytesProcessed();
-}
-
-function rtOnceReadBits(f, view) {
-  const rs = new ReadStream(view);
-  return readRtBits(rs, f);
-}
-
-function rtOnceWriteMixed(f, buf) {
-  const ws = new WriteStream(buf);
-  if (!writeRtMixed(ws, f)) {
-    return -1;
-  }
-  ws.flush();
-  return ws.bytesProcessed();
-}
-
-function rtOnceReadMixed(f, view) {
-  const rs = new ReadStream(view);
-  return readRtMixed(rs, f);
-}
-
-// ---- the timed loops, one function per (shape, path) as in the Go and C#
-// runners; streams reuse via reset — the runtime's documented no-allocation
-// path, as the gen benches do ----
-
-function rtBenchPacketWriteLoop(base, iters) {
-  const ws = new WriteStream(gBuffer);
-  for (let i = 0; i < iters; i++) {
-    lcgStep();
-    varyRtPacket(base);
-    ws.reset(gBuffer);
-    if (!writeRtPacket(ws, base)) {
-      return false;
-    }
-    ws.flush();
-    gSink = (gSink + ws.bytesProcessed()) >>> 0;
-  }
-  return true;
-}
-
-function rtBenchPacketReadLoop(outValue, iters, views) {
-  const rs = new ReadStream(views[0]);
-  for (let i = 0; i < iters; i++) {
-    rs.reset(views[i & (NumVariants - 1)]);
-    if (!readRtPacket(rs, outValue)) {
-      return false;
-    }
-    gSink = (gSink + sinkOfRtPacket(outValue)) >>> 0; // full-struct observation (#175)
-  }
-  return true;
-}
-
-function rtBenchIntsWriteLoop(base, iters) {
-  const ws = new WriteStream(gBuffer);
-  for (let i = 0; i < iters; i++) {
-    lcgStep();
-    varyRtInts(base);
-    ws.reset(gBuffer);
-    if (!writeRtInts(ws, base)) {
-      return false;
-    }
-    ws.flush();
-    gSink = (gSink + ws.bytesProcessed()) >>> 0;
-  }
-  return true;
-}
-
-function rtBenchIntsReadLoop(outValue, iters, views) {
-  const rs = new ReadStream(views[0]);
-  for (let i = 0; i < iters; i++) {
-    rs.reset(views[i & (NumVariants - 1)]);
-    if (!readRtInts(rs, outValue)) {
-      return false;
-    }
-    gSink = (gSink + sinkOfRtInts(outValue)) >>> 0; // full-struct observation (#175)
-  }
-  return true;
-}
-
-function rtBenchBitsWriteLoop(base, iters) {
-  const ws = new WriteStream(gBuffer);
-  for (let i = 0; i < iters; i++) {
-    lcgStep();
-    varyRtBits(base);
-    ws.reset(gBuffer);
-    if (!writeRtBits(ws, base)) {
-      return false;
-    }
-    ws.flush();
-    gSink = (gSink + ws.bytesProcessed()) >>> 0;
-  }
-  return true;
-}
-
-function rtBenchBitsReadLoop(outValue, iters, views) {
-  const rs = new ReadStream(views[0]);
-  for (let i = 0; i < iters; i++) {
-    rs.reset(views[i & (NumVariants - 1)]);
-    if (!readRtBits(rs, outValue)) {
-      return false;
-    }
-    gSink = (gSink + sinkOfRtBits(outValue)) >>> 0; // full-struct observation (#175)
-  }
-  return true;
-}
-
-function rtBenchMixedWriteLoop(base, iters) {
-  const ws = new WriteStream(gBuffer);
-  for (let i = 0; i < iters; i++) {
-    lcgStep();
-    varyRtMixed(base);
-    ws.reset(gBuffer);
-    if (!writeRtMixed(ws, base)) {
-      return false;
-    }
-    ws.flush();
-    gSink = (gSink + ws.bytesProcessed()) >>> 0;
-  }
-  return true;
-}
-
-function rtBenchMixedReadLoop(outValue, iters, views) {
-  const rs = new ReadStream(views[0]);
-  for (let i = 0; i < iters; i++) {
-    rs.reset(views[i & (NumVariants - 1)]);
-    if (!readRtMixed(rs, outValue)) {
-      return false;
-    }
-    gSink = (gSink + sinkOfRtMixed(outValue)) >>> 0; // full-struct observation (#175)
-  }
-  return true;
-}
-
-// ---- the family rt driver: §1.5 oracle gate, then the timed loops ----
-
-function benchRt(name, iters, base, outValue, onceWrite, onceRead, writeLoop, readLoop, vary, sinkOf) {
-  // oracle 1: the hand-written wire must equal the generated-code golden
-  const bytesPerOp = onceWrite(base, gBuffer);
-  if (bytesPerOp < 0) {
-    fail(name, "write of pinned instance failed");
-    return;
-  }
-  if (!checkGolden(name, gBuffer.subarray(0, bytesPerOp))) {
-    failed = true;
-    return;
-  }
-
-  // oracle 2: round-trip write -> read -> re-write -> identical bytes
-  if (!onceRead(outValue, gBuffer.subarray(0, bytesPerOp))) {
-    fail(name, "read of pinned instance failed");
-    return;
-  }
-  if (onceWrite(outValue, gTwin) !== bytesPerOp || !bytesEqual(gTwin.subarray(0, bytesPerOp), gBuffer.subarray(0, bytesPerOp))) {
-    fail(name, "round-trip bytes differ");
-    return;
-  }
-  if (!Number.isFinite(sinkOf(outValue))) {
-    fail(name, "sinkOf not finite on the decoded pinned instance (typo'd field?)");
-    return;
-  }
-
-  // variant buffers (and proof that variation keeps bytes/op constant)
-  lcgSeed(1);
-  const views = [];
-  for (let k = 0; k < NumVariants; k++) {
-    lcgStep();
-    vary(base);
-    if (onceWrite(base, gVariants[k]) !== bytesPerOp) {
-      fail(name, "variation changed bytes/op — vary must keep structure fields fixed");
-      return;
-    }
-    views.push(gVariants[k].subarray(0, bytesPerOp));
-  }
-
-  const writeRates = new Array(gNumRuns);
-  const readRates = new Array(gNumRuns);
-
-  for (let run = -1; run < gNumRuns; run++) {
-    const start = performance.now();
-    if (!writeLoop(base, iters)) {
-      fail(name, "write failed in loop");
-      return;
-    }
-    const elapsed = (performance.now() - start) / 1000.0;
-    if (run >= 0) {
-      writeRates[run] = iters / elapsed;
-    }
-  }
-
-  for (let run = -1; run < gNumRuns; run++) {
-    const start = performance.now();
-    if (!readLoop(outValue, iters, views)) {
-      fail(name, "read failed in loop");
-      return;
-    }
-    const elapsed = (performance.now() - start) / 1000.0;
-    if (run >= 0) {
-      readRates[run] = iters / elapsed;
-    }
-  }
-
-  report(name, "write", iters, bytesPerOp, stats(writeRates), "rt");
-  report(name, "read", iters, bytesPerOp, stats(readRates), "rt");
-}
-
-// --------------------------------------------------------------------------
 // family bits (BENCH-STANDARD.md §1.4): the raw BitWriter/BitReader with
 // the 16-width table (227 bits/group) over a 65536-byte buffer, the ONE
 // bitpacker workload in the estate. Values vary per pass through the LCG
@@ -2197,12 +1386,9 @@ function benchBitpacker(passes) {
 }
 
 // --------------------------------------------------------------------------
-
-
-// --------------------------------------------------------------------------
 // the four Bench.schema shapes as GENERATED code (flat tier = THE js path):
 // pins and vary functions over the generated classes, field mappings
-// verbatim from the family benches; the same LCG draws as the rt legs
+// verbatim from the family benches
 // --------------------------------------------------------------------------
 
 function pinBenchPacketGen() {
@@ -2384,21 +1570,11 @@ function main() {
   // the four Bench.schema shapes as GENERATED code — the flat tier is THE
   // js entry for these shapes in any cross-language comparison, exactly as
   // for the corpus shapes above (family gen, codec=flat), cross-validated
-  // against the runtime-call tier by the same oracle. The family rt rows
-  // below keep the same bench names and measure the serialize.js runtime
-  // API instead — honest library-context data, never the js number.
+  // against the runtime-call tier by the same oracle.
   benchMessageFlat("bench_packet", "bench_packet", 32000000, pinBenchPacketGen(), bench.WriteBenchPacket, bench.ReadBenchPacket, benchFlat.WriteBenchPacketFlat, benchFlat.ReadBenchPacketFlat, varyBenchPacketGen, sinkOfBenchPacketGen);
   benchMessageFlat("bench_ints", "bench_ints", 40000000, pinBenchIntsGen(), bench.WriteBenchInts, bench.ReadBenchInts, benchFlat.WriteBenchIntsFlat, benchFlat.ReadBenchIntsFlat, varyBenchIntsGen, sinkOfBenchIntsGen);
   benchMessageFlat("bench_bits", "bench_bits", 48000000, pinBenchBitsGen(), bench.WriteBenchBits, bench.ReadBenchBits, benchFlat.WriteBenchBitsFlat, benchFlat.ReadBenchBitsFlat, varyBenchBitsGen, sinkOfBenchBitsGen);
   benchDataDrivenFlat("bench_mixed", "bench_mixed", 4000000, bench.BenchMixed, bench.WriteBenchMixed, bench.ReadBenchMixed, benchFlat.WriteBenchMixedFlat, benchFlat.ReadBenchMixedFlat);
-
-  // family rt (§1.3/§1.5): the runtime API by hand, oracle-gated against
-  // the goldens the generated code pinned. Iteration counts are fixed and
-  // identical across all six runners (§2.1; sized in the C++ reference).
-  benchRt("bench_packet", 32000000, pinRtPacket(), makeRtPacket(), rtOnceWritePacket, rtOnceReadPacket, rtBenchPacketWriteLoop, rtBenchPacketReadLoop, varyRtPacket, sinkOfRtPacket);
-  benchRt("bench_ints", 40000000, pinRtInts(), makeRtInts(), rtOnceWriteInts, rtOnceReadInts, rtBenchIntsWriteLoop, rtBenchIntsReadLoop, varyRtInts, sinkOfRtInts);
-  benchRt("bench_bits", 48000000, pinRtBits(), makeRtBits(), rtOnceWriteBits, rtOnceReadBits, rtBenchBitsWriteLoop, rtBenchBitsReadLoop, varyRtBits, sinkOfRtBits);
-  benchRt("bench_mixed", 4000000, pinRtMixed(), makeRtMixed(), rtOnceWriteMixed, rtOnceReadMixed, rtBenchMixedWriteLoop, rtBenchMixedReadLoop, varyRtMixed, sinkOfRtMixed);
 
   // family bits (§1.4): the one bitpacker workload in the estate
   benchBitpacker(24576);
